move e2e_node_slice to its own package and import into e2e tests
