+++ e2e/e2e_test.go
@@ -26,6 +26,9 @@ import (
 	"github.com/k8snetworkplumbingwg/whereabouts/pkg/api/whereabouts.cni.cncf.io/v1alpha1"
 	wbstorage "github.com/k8snetworkplumbingwg/whereabouts/pkg/storage/kubernetes"
 	"github.com/k8snetworkplumbingwg/whereabouts/pkg/types"
+
+	// Import node slice tests to execute in the suite
+	_ "github.com/k8snetworkplumbingwg/whereabouts/e2e/e2e_node_slice"
 )
 
 func TestWhereaboutsE2E(t *testing.T) {
